% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_rmse.R
\name{frac_rmse}
\alias{frac_rmse}
\alias{frac_rmse.data.frame}
\alias{frac_rmse_vec}
\title{partial root mean square error}
\usage{
frac_rmse(data, ...)

\method{frac_rmse}{data.frame}(
  data,
  truth,
  estimate,
  frac = 0.1,
  ans_order = "desc",
  na_rm = TRUE,
  case_weights = NULL,
  ...
)

frac_rmse_vec(
  truth,
  estimate,
  frac = 0.1,
  ans_order = "desc",
  na_rm = TRUE,
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A `data.frame` containing the columns specified by the `truth`
and `estimate` arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results
(that is `numeric`). This should be an unquoted column name although
this argument is passed by expression and supports
[quasiquotation][rlang::quasiquotation] (you can unquote column
names). For `_vec()` functions, a `numeric` vector.}

\item{estimate}{The column identifier for the predicted
results (that is also `numeric`). As with `truth` this can be
specified different ways but the primary method is to use an
unquoted variable name. For `_vec()` functions, a `numeric` vector.}

\item{frac}{Proportion of data to use in calculating rmse. If the value is less then 1,
then it is considered as a proportion. If the value is greater than or equal to 1,
then it is considered as a number of data points used in calculation.}

\item{na_rm}{A `logical` value indicating whether `NA`
values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights. This
should be an unquoted column name that evaluates to a numeric column in
`data`. For `_vec()` functions, a numeric vector.}
}
\description{
partial root mean square error
}
