% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_dat.R
\name{get_plot_dat}
\alias{get_plot_dat}
\title{포트폴리오 시각화를 위한 데이터 생성}
\usage{
get_plot_dat(port_return, bm_return)
}
\arguments{
\item{port_return}{포트폴리오 일간 수익률. rtn_with_cost 변수가 열에 있어야 함.}

\item{bm_return}{벤치마크 일간 수익률}
}
\value{
type에 포트/bm이 표시된 long format 누적 수익률
}
\description{
포트 일간 수익률 데이터로부터 누적 수익률 데이터 생성
포트 수익률은
포트/bm 수익률 데이터가 모두 존재하는 기간에 대해서 수익률 생성
}
