% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_sim.R
\name{get_port_sim}
\alias{get_port_sim}
\title{포트폴리오 시뮬 summary statistic 생성}
\usage{
get_port_sim(port_return)
}
\arguments{
\item{port_return}{get_port_return 함수에서 반환된 포트폴리오 수익률}
}
\value{
시뮬레이션 결과 요약 tbl
}
\description{
port_return으로부터 시뮬레이션 결과 생성
연간수익률, 최대낙폭, 평균 턴오버 계산
}
