% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{step_winsorize}
\alias{step_winsorize}
\title{winsorize 레시피 스텝}
\usage{
step_winsorize(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  probs = c(0.05, 0.95),
  min_vals = NULL,
  max_vals = NULL,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("winsorize")
)
}
\arguments{
\item{recipe}{recipe object}

\item{role}{}
}
\description{
tidy modeling시 사용할 레시피 스텝
}
