% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_function.R
\name{tbl_qpms}
\alias{tbl_qpms}
\alias{connect_qpms_db}
\alias{disconnect_qpms_db}
\title{DB 연결 함수}
\usage{
tbl_qpms(dbo_name, table_name)

connect_qpms_db(dsn = "QPMS", encoding = "EUC-KR")

disconnect_qpms_db()
}
\arguments{
\item{dbo_name}{dbo 이름}

\item{table_name}{테이블 명}
}
\value{
collect 전의 tbl
}
\description{
QPMS DB에 연결하여 데이터 반환
}
