% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_return.R
\name{get_port_return}
\alias{get_port_return}
\title{포트폴리오 일간 수익률 계산 함수}
\usage{
get_port_return(port_weight, rtn_tbl, trd_cost = 0.003)
}
\arguments{
\item{port_weight}{포트비중 시계열. td(날짜), ticker(티커명), weight(비중) 열을 가지고 있어야 함.}

\item{rtn_tbl}{종목 일간 수익률. 종목에 대한 일간수익률 데이터는
td, ticker, rtn(수익률) 열을 가지고 있어야 함.}

\item{trd_cost}{거래비용. 디폴트는 30bp}
}
\value{
일별 포트 수익률과 비용감안 포트 수익률률
}
\description{
포트폴리오 비중에 대해 거래수수료를 감안한 일간 수익률 계산
포트비중의 날짜 종가로
}
